requirements.txt
bcrypt==5.0.0
blinker==1.8.2
certifi==2025.10.5
cffi==1.17.1
charset-normalizer==3.4.4
click==8.1.8
cryptography==46.0.3
flask==3.0.3
future==1.0.0
idna==3.11
importlib-metadata==8.5.0
itsdangerous==2.2.0
jinja2==3.1.6
MarkupSafe==2.1.5
netmiko==4.4.0
ntc-templates==7.9.0
paramiko==3.5.1
pycparser==2.23
PyJWT==1.7.1
pynacl==1.6.0
pyserial==3.5
python-dotenv==1.0.1
PyYAML==6.0.3
requests==2.32.4
requests-toolbelt==1.0.0
scp==0.15.0
six==1.17.0
ssh-python==1.2.0.post1
textfsm==1.1.3
typing-extensions==4.13.2
urllib3==2.2.3
webexteamssdk==1.7
werkzeug==3.0.6
zipp==3.20.2
